added extra date + permissions check

+++ mod_shoutbox/mod_shoutbox.php
@@ -80,6 +80,10 @@ elseif ($date == 4)
 {
 	$show_date = "D j M - ";
 }
+elseif ($date == 5)
+{
+	$show_date = "Y/m/d - ";
+}
 else
 {
 	$show_date = "";
@@ -120,7 +124,10 @@ if (isset($_POST))
 	if (isset($post['delete']))
 	{
 		$deletepostnumber = $post['idvalue'];
-		ModShoutboxHelper::deletepost($deletepostnumber);
+		if ($user->authorise('core.delete'))
+		{
+			ModShoutboxHelper::deletepost($deletepostnumber);
+		}
 	}
 
 	if ($mass_delete == 0)
@@ -139,8 +146,10 @@ if (isset($_POST))
 						{
 							$delete = $post['max'];
 						}
-
-						modShoutboxHelper::deleteall($delete);
+						if ($user->authorise('core.delete'))
+						{
+							modShoutboxHelper::deleteall($delete);
+						}
 					}
 					else
 					{
